Add 'quiet' param for LoadPlugins
Plugins use a different logger form the rest of the framework due to the
fact that they run in a different process. Add a 'quiet' flag to
LoadPlugins (and expose it via appropriate InializeStore() methods) to
allow suppressing output from the plugin.

(Note such output can always be considered to be "debug", as any actual
issue will be propagated as errors to the main process).

This is primarily to allow suppressing of intermittent output for
testing.