common/param: fix populate map from Viper
Inside ParamStore.PopulateFromViper(), values identified as reflect.Map
were being retrieved as map[string]string, which is an invalid structpb
value type. This is a problem as param data is stored as a
structpb.Struct so that it can be conveyed via a protobuf channel.
Retrieve reflect.Map's as map[string]interface{} instead.

Signed-off-by: Sergei Trofimov <Sergei.Trofimov@arm.com>